refactor: Move PlacementPopupPanel functionality to base panel
